// Exercise_Fourteen/How-to-connect-mongose-in-typescript/src/Types/Post.types.ts
interface PostData {
  title: string;
  content: string;
  author: string;
}

export interface IGetAllPostsResponse {
  status: string;
  message: string;
  currentPosts: PostData[];
}
export interface ICreatePostResponse {
  status: string;
  message: string;
  currentPosts: {
    title: string;
    content: string;
    author: string;
  };
}